explore implementing api to model mapping across modules
A fundamental part of the separation of concerns in the architecture of the app is to be able to define an application core model, which becomes the internal contract for working with objects through out the application, along with a clean and simple definition of api json entities and the ability to quickly map/transform these entities to model entities. This work is about exploring and refining how this flow works when the Kotlin defined entities are contained within separate application modules.

 Need to refine the definitions/imports to avoid the shotgun approach that it has taken to get this to work. Also work through implementing constructors for the data classes that allow a clean simple empty constructor instance creation.